refactor: :ambulance: refactor image redactor
